requirements.txt
attrs==21.4.0
bcrypt==3.2.0
cffi==1.15.0
click==8.0.3
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.2
idna==3.3
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pydantic==1.8.2
PyJWT==2.3.0
pyparsing==3.0.6
pytest==6.2.5
six==1.16.0
toml==0.10.2
typing-extensions==4.0.1
Werkzeug==2.0.2
